requirements.txt
streamlit
pillow
opencv-python-headless
numpy
torch
matplotlib
